feat: add spa starter kit content and make UX wireframe
Start off Archimedes with the spa-starter-kit salvaged from Euclid,
there is still a bit to do on that to clean things up. However, I wanted
to get what I have in here now checked in to have a good foundation to
continue working on this.

In addition, I have created a first draft of the UX wireframe for
Archimedes. It is still a bit rough, and missing things: such as
the feed editing screen, but it does a pretty good job of representing
what I am thinking about making Archimedes into.

Signed-off-by: David Piper <david.piper@mailfence.com>