docs(wireframe): add the edit feed page to wireframe
Add the Edit Feed page to the Wireframe design, this page is needed
because the user will have to have a way to create new feeds (since
there will be none at the start) and to edit existing feeds. Editing
means: changing the title of the feed, adding new sources to watch,
and removing existing sources that are being watched.

This also includes the feature to import sources from package.json files
which is a feature that I have been planning, but hadn't yet documented
in an official manner.

Signed-off-by: David Piper <david.piper@mailfence.com>