feat(server): add GraphQL API to server for releases
Add a GraphQL API to the Server for accessing Github releases. This
is basically just a wrapper around the Github GraphQL API, but it is
good to be doing it this way; rather than accessing it directly on
the client. So that we don't have to worry about CORS and also so
that we have a centralized interface for accessing this data; which
may be different for other sources.

As an example: Gitlab might have a completely different API, but we
can wrap it the exact same way and the client won't have to know about
the difference.

Also fixed a minor annoyance with the layout of the Wireframe; the
html view in Pencil has very strange spacing between lines. I simulated
it by using multiple labels, and was able to get the lines closer
together which looks much better.

Signed-off-by: David Piper <david.piper@mailfence.com>