feat(client): build a basic prototype of the client
Build a basic prototype of the client, I still need to do a
real UX prototype design; however this is a quick example of
what the client can do. I thought it would be good to actually
have the client fetch the server data and display it for this
issue before going off and doing anymore UX design.

Signed-off-by: David Piper <david.piper@mailfence.com>