docs: replace all references to Euclid with Archimedes
Since this project was bootstrapped with the remnants of Euclid
there were many erroneous references to Euclid scattered throughout
the project. This replaces all of them with the name of this
project: Archimedes. It also writes a basic readme for Archimedes
explaining what it is and the motivation for it.

Signed-off-by: David Piper <david.piper@mailfence.com>