docs(mockup): add release feed content section
Add the release feed content section to the mockup, also
add a footer where I give credit to creative commons authors'
work; for example for the edit icon. Come up with a better
positioning for the release feed title and the date picker
drop-down, and used an edit (pencil) icon instead of the
word "Edit".

In addition I cleaned up the styling of the various text elements,
and started using the "Atomic Design" methodology to keep things
nice and DRY.

Signed-off-by: David Piper <david.piper@mailfence.com>