docs(mockup): finish up mockup
Finish up the initial Archimedes UI Design mockup, specifically
created the Feed Cards on the Home screen and also created the
Edit screen. Some free, Creative Commons, icons are used in the
UI Design and these need to be properly attributed on the
Front-end; as done in the UI Mockup.

Signed-off-by: David Piper <david.piper@mailfence.com>