feat(client): create H1 atom for front-end based on UI mockup
I am going to follow the atomic design layout for structuring the
front end as well as the design, as I really like it and think that
it is a pretty good approach. It automatically enforces the
components to be as simple as possible, thus making the React
front-end code well-structured and easy to navigate.

The previous directory structure was way too rigid and made things
incredibly confusing, since it was based on screen and digging down
tons of layers. I think that atomic design will have many of the
same benefits, without many of the problems and will also have
the benefits of a "flat" directory structure. Thus, bringing the
best of both worlds!

Signed-off-by: David Piper <david.piper@mailfence.com>