client(atoms): create H1-Subtitle2 text components
Also had to tweak the design of the mockup a bit due
to Google Fonts not having a medium font weight of Lato
(though Figma said it did). I came up with a hack where
I increase the font size using the Regular weight by 1/140.
This results in a very slight increase in size/weight which
should achieve the desired result.

Signed-off-by: David Piper <david.piper@mailfence.com>