feat(client): add several new simple molecules
Furthermore, I ended up refactoring the Switches and
Checkbox to add functionality such as: on/off for Switches,
onClick events, and using className. This makes the switches and
checkbox components behave very similarly to the native HTML5
checkboxes which makes for a better DX as the developers
will find that it is more familiar to them.

Signed-off-by: David Piper <david.piper@mailfence.com>