feat(client): make VerticalVoteButtons component
In addition, refactored the tests to use the names from the
React components directly rather than relying on typing them
in as a string. This reduces potential for error and makes it
more automated in case of things like renaming components.

Also took more steps to further speed up the Visual Regression
tests by directly loading the iframes, as it was mistakenly
vising the main Storybook page still.

Signed-off-by: David Piper <david.piper@mailfence.com>