test(client): test the voting functionality of VerticalVoteButton
This required refactoring the way that animated tests are done as it
needed to test two properties on two different elements, each with the
same regex, but differing indices! Which meant that either a ton of
engineering needed to be done to test this (tried this but it was pretty
crazy) _or_ just make the utilities a bit simpler, but more powerful and
flexible. So it does the latter.

Signed-off-by: David Piper <david.piper@mailfence.com>