refactor(client): use new aliases: @src, @util, @atoms, and @molecules
I had the idea to start using more aliases to improve the readability
and usability of imports. This is especially helpful if you _know_
that you want something and that it is an atom, but you are making
a molecule so you have to type `../../../atoms/{myComponent}` this
is silly and can easily be avoided by having a @atoms alias, overall
this should dramatically help with these cases.

Signed-off-by: David Piper <david.piper@mailfence.com>