fix: convert the remaining TSLint rules to use relative paths, fixes CI
The CI, and any other system other than mine, was failing because the
TSLint paths were absolute paths on my system. I guess that this is the
default behavior of the tslint-to-eslint-config tool, and it got missed
with local tests because they _did_ work on my system; just not on the
CI. This is a great example of why _both_ local tests and CI are needed
to find potential issues as early as possible!

Signed-off-by: David Piper <david.piper@mailfence.com>